re-set default player to async
Also improve Safari compatibility under proxy-to-pthread condition
