+++ Makefile.emscripten
@@ -54,8 +54,7 @@ HAVE_7ZIP = 1
 HAVE_BSV_MOVIE = 1
 HAVE_AL = 1
 HAVE_CHD ?= 0
-HAVE_WASMFS ?= 1
-HAVE_WORKER ?= 1
+HAVE_WASMFS ?= 0
 PROXY_TO_PTHREAD ?= 0
 HAVE_NETPLAYDISCOVERY ?= 0
 
@@ -74,14 +73,14 @@ HAVE_RWEBAUDIO = 0
 GL_DEBUG ?= 0
 
 # enable javascript filesystem tracking
-FS_DEBUG = 1
+FS_DEBUG = 0
 
 HAVE_OPENGLES ?= 1
 HAVE_OPENGLES3 ?= 0
 
 ASYNC ?= 1
 LTO ?= 0
-PTHREAD ?= 4
+PTHREAD ?= 0
 
 STACK_SIZE ?= 4194304
 INITIAL_HEAP ?= 134217728
@@ -110,15 +109,13 @@ EXPORTS := callMain,FS,PATH,ERRNO_CODES,ENV,stringToNewUTF8,UTF8ToString,Browser
 LIBS    := -s USE_ZLIB=1 -lbrowser.js
 
 ifeq ($(HAVE_WASMFS), 1)
-   LIBS += -s WASMFS -s FORCE_FILESYSTEM=1 -lfetchfs.js -lopfs.js
+   DEFINES += -DHAVE_WASMFS=1
+   LIBS += -sWASMFS -sFORCE_FILESYSTEM=1 -lfetchfs.js -lopfs.js
    EXPORTS := $(EXPORTS),FETCHFS,OPFS
 endif
 
-ifeq ($(HAVE_WORKER), 1)
-   LIBS += -s USE_ES6_IMPORT_META=0 -sENVIRONMENT=worker,web
-endif
-
 ifeq ($(PROXY_TO_PTHREAD),1)
+   LIBS += -sUSE_ES6_IMPORT_META=0 -sENVIRONMENT=worker,web
    LIBS += -sPROXY_TO_PTHREAD -sOFFSCREENCANVAS_SUPPORT
    DEFINES += -DUSE_OFFSCREENCANVAS=1 -DPROXY_TO_PTHREAD=1
 else
