+++ pkg/emscripten/libretro-thread/embed.html
