+++ pkg/emscripten/libretro/index.html
@@ -197,7 +197,8 @@
       <script crossorigin="anonymous" src="//cdnjs.cloudflare.com/ajax/libs/tether/1.3.4/js/tether.min.js"></script>
       <script crossorigin="anonymous" src="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-alpha.3/js/bootstrap.min.js"></script>
       <script src="analytics.js"></script>
-      <script src="zip-no-worker.min.js"></script>
+      <!--script src="//wzrd.in/standalone/browserfs@0.6.1"></script-->
+      <script src="browserfs.min.js"></script>
       <script src="libretro.js"></script>
    </body>
 </html>
